Inline the remaining coding bits into `Codec`.
